+++ neo/modules/cse/__init__.py
@@ -1,3 +1,4 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (C) 2023 sardonicism-04
+from .objects import GoogleResponse, SearchResult
 from .search import Search